+++ tests/test_emitters.py
@@ -12,7 +12,7 @@ from aioway.emits import MlpEmitter, emit, linear_shape
 from aioway.io import TensorListIter, TensorStream
 from aioway.relalg import LoaderOpt
 from aioway.spaces import AttrSpace, ShapeSpace
-from aioway.torch.nn import Linear, MSELoss, NnUFunc
+from aioway.torch.nn_ import Linear, MSELoss, NnUFunc
 from aioway.torch.optim import OptimizerUFunc
