+++ tests/torch/nn/test_nn_init.py
@@ -4,7 +4,7 @@
 from torch import nn
 
 from aioway.modes import NnInitThunk
-from aioway.torch.nn import NnInit_, Sequential, find_nn_init
+from aioway.torch.nn_ import NnInit_, Sequential, find_nn_init
 
 
 def test_layer_nn_init(layer_nn_init: NnInit_):
